Wait condition improvements (#2)
* Fix referential transparency of WaitForConstraint
* Add WaitForConditionOn, to allow cleaner, more optimized code
* Change seconds from int to double, this should be practically backwards compatible
* Add WaitForFrames

+++ Packages/Responsible/Runtime/TestWaitCondition.cs
@@ -128,7 +128,7 @@ namespace Responsible
 		[Pure]
 		public static ITestInstruction<T> ExpectWithinSeconds<T>(
 			this ITestWaitCondition<T> condition,
-			int timeout,
+			double timeout,
 			[CallerMemberName] string memberName = "",
 			[CallerFilePath] string sourceFilePath = "",
 			[CallerLineNumber] int sourceLineNumber = 0)
